refactor(auth): move password handlers out of oauth_providers.go into dedicated password.go
