+++ internal/features/auth/handlers/validate.go
@@ -5,6 +5,11 @@ import (
 	"net/mail"
 	"strings"
 	"unicode"
+
+	"github.com/danielgtaylor/huma/v2"
+	"github.com/google/uuid"
+
+	"github.com/shuvo-paul/medminder/internal/features/auth/service"
 )
 
 var (
@@ -74,4 +79,32 @@ func ValidateDisplayName(name string) error {
 		return ErrInvalidDisplayName
 	}
 	return nil
+}
+
+func ExtractUserIDFromAuth(authHeader string, tokenSvc service.TokenServiceInterface) (uuid.UUID, error) {
+	if len(authHeader) < 7 || authHeader[:7] != "Bearer " {
+		return uuid.Nil, huma.Error401Unauthorized("Invalid authorization header", nil)
+	}
+	tokenString := authHeader[7:]
+
+	claims, err := tokenSvc.ValidateAccessToken(tokenString)
+	if err != nil {
+		return uuid.Nil, huma.Error401Unauthorized("Invalid or expired access token", err)
+	}
+
+	userIDStr, ok := claims["sub"].(string)
+	if !ok {
+		return uuid.Nil, huma.Error401Unauthorized("Invalid access token", nil)
+	}
+
+	userID, err := uuid.Parse(userIDStr)
+	if err != nil {
+		return uuid.Nil, huma.Error401Unauthorized("Invalid user ID in token", nil)
+	}
+
+	if userID == uuid.Nil {
+		return uuid.Nil, huma.Error401Unauthorized("Invalid user ID", nil)
+	}
+
+	return userID, nil
 }
